:rocket: Splitted API in separated files
