+++ pyproject.toml
@@ -7,14 +7,17 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.10"
-pytest = "^7.2.0"
 fastapi = "^0.88.0"
 uvicorn = "^0.20.0"
 python-dotenv = "^0.21.0"
 motor = "^3.1.1"
-pytest-asyncio = "^0.20.3"
-flaky = "^3.7.0"
 
+
+[tool.poetry.group.test.dependencies]
+flaky = "^3.7.0"
+pytest = "^7.2.0"
+behave = "^1.2.6"
+pytest-asyncio = "^0.20.3"
 
 [build-system]
 requires = ["poetry-core"]
