+++ tests/Counter/application/test_counter_finder.py
@@ -1,17 +1,20 @@
 from ..domain import CounterStub
-from ..infrastructure import FakeCounterRepository
+from ..fakes import FakeCounterRepository
+
+from src.contexts.Counter.domain.repositories import CounterRepository
 
 from Counter.application.find import CounterFinder
 
 import pytest
+
 
 class TestCounterFinder:
     @pytest.mark.asyncio
     async def test_can_find_counter(self):
         # given
         counter = CounterStub.random()
-
         repo = FakeCounterRepository({counter})
+
         finder = CounterFinder(repo)
 
         # when
