+++ tests/Counter/fakes/FakeCounterRepository.py
