+++ tests/apps/backend/app/features/steps/create_counter.py
@@ -1,0 +1,17 @@
+from behave import *
+
+@given('an user with id {userId}')
+def step_impl(context):
+    pass
+
+@given('a {private} counter with id {counterId}')
+def step_impl(context):
+    pass
+
+@when('the user tries to create the counter')
+def step_impl(context):
+    pass
+
+@then('the counter is createt as desired')
+def step_impl(context):
+    pass
